backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
django-environ==0.11.2
dj-database-url==2.1.0
django-filter==24.3
django-debug-toolbar==4.4.2
psycopg2-binary==2.9.9
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
Pillow==10.4.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
